Include camera in scene
