+++ scenes/bokeh_demo.yml
@@ -1,3 +1,16 @@
+camera:
+  image_width: 720
+  image_height: 480
+  location: { x: 0.0, y: -0.01, z: -15.0 }
+  yaw: 0.0
+  pitch: 0.0
+  roll: 0.0
+  sensor_width: 0.036
+  sensor_height: 0.024
+  focal_length: 0.05
+  focus_distance: 1.0
+  aperture: 2.0
+
 skybox:
   type: Gradient
   overhead_colour: { r: 0.5, g: 0.8, b: 0.9 }
