+++ src/paths/renderer.rs
@@ -4,24 +4,23 @@ use rand;
 use rand::Rng;
 use threadpool::ThreadPool;
 
-use crate::paths::{Camera, Image, Ray};
+use crate::paths::{Image, Ray};
 use crate::paths::colour::{Colour};
 use crate::paths::pixels::Estimator;
 use crate::paths::scene::Scene;
 use crate::paths::vector::Vector3;
 
 pub struct Renderer {
-    scene: Scene,
-    pub camera: Camera,
+    pub scene: Scene,
     estimator: Estimator,
     pool: ThreadPool,
 }
 
 impl Renderer {
-    pub fn new(scene: Scene, camera: Camera, num_workers: usize) -> Renderer {
-        let estimator = Estimator::new(camera.width as usize, camera.height as usize);
+    pub fn new(scene: Scene, num_workers: usize) -> Renderer {
+        let estimator = Estimator::new(scene.camera.width as usize, scene.camera.height as usize);
         let pool = ThreadPool::new(num_workers);
-        Renderer{ scene, camera, estimator, pool}
+        Renderer{ scene, estimator, pool}
     }
 
     pub fn render(&self) -> Image {
@@ -31,12 +30,12 @@ impl Renderer {
     pub fn trace_full_pass(&mut self) {
         let (tx, rx) = channel::<(u32, u32, Colour)>();
 
-        self.camera.init_bundle();
+        self.scene.camera.init_bundle();
 
-        for x in 0 .. self.camera.width {
+        for x in 0 .. self.scene.camera.width {
             let tx = tx.clone();
             let scene = self.scene.clone();
-            let mut camera = self.camera.clone();
+            let mut camera = self.scene.camera.clone();
 
             self.pool.execute(move|| {
                 for y in 0 .. camera.height {
@@ -48,7 +47,7 @@ impl Renderer {
             });
         }
 
-        let num_pixels = self.camera.height * self.camera.width;
+        let num_pixels = self.scene.camera.height * self.scene.camera.width;
         rx.iter().take(num_pixels as usize).for_each(|(x, y, colour)| {
             self.estimator.update_pixel(x as usize, y as usize, colour);
         });
@@ -57,14 +56,14 @@ impl Renderer {
     pub fn trace_batch(&mut self, num_rays: usize) {
         for _ in 0 .. num_rays {
             let (x, y) = self.estimator.choose_pixel();
-            let ray = self.camera.get_ray_for_pixel(x as u32, y as u32);
+            let ray = self.scene.camera.get_ray_for_pixel(x as u32, y as u32);
             let colour = Renderer::trace_ray(&self.scene, ray, 0);
             self.estimator.update_pixel(x as usize, y as usize, colour);
         }
     }
 
     pub fn reset(&mut self) {
-        self.estimator = Estimator::new(self.camera.width as usize, self.camera.height as usize);
+        self.estimator = Estimator::new(self.scene.camera.width as usize, self.scene.camera.height as usize);
     }
 
     fn trace_ray(scene: &Scene, ray: Ray, depth: u32) -> Colour {
